AULAS/2-SEMESTRE/LNP/requirements.txt
Flask==2.3.3
google-cloud-dialogflow==2.24.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0
pandas==2.1.1
python-dotenv==1.0.0
gunicorn==21.2.0
flask-cors==4.0.0
